+++ src/utils/xml/SUMOSAXAttributesImpl_Cached.cpp
@@ -66,7 +66,7 @@ SUMOSAXAttributesImpl_Cached::~SUMOSAXAttributesImpl_Cached() { }
 bool
 SUMOSAXAttributesImpl_Cached::hasAttribute(int id) const {
     assert(id >= 0); 
-    assert(id < myPredefinedTagsMML.size());
+    assert(id < (int)myPredefinedTagsMML.size());
     return myAttrs.find(myPredefinedTagsMML[id]) != myAttrs.end();
 }
 
@@ -111,7 +111,7 @@ SUMOSAXAttributesImpl_Cached::getFloat(int id) const {
 const std::string&
 SUMOSAXAttributesImpl_Cached::getAttributeValueSecure(int id) const {
     assert(id >= 0); 
-    assert(id < myPredefinedTagsMML.size());
+    assert(id < (int)myPredefinedTagsMML.size());
     return myAttrs.find(myPredefinedTagsMML[id])->second;
 }
 
@@ -236,7 +236,7 @@ SUMOSAXAttributesImpl_Cached::getBoundary(int attr) const {
 std::string
 SUMOSAXAttributesImpl_Cached::getName(int attr) const {
     assert(attr >= 0); 
-    assert(attr < myPredefinedTagsMML.size());
+    assert(attr < (int)myPredefinedTagsMML.size());
     return myPredefinedTagsMML[attr];
 }
